changed classdef generation handling of WSDL simpleType enumeration: it was converted to String constants in a module.  now it generates constants of a class as follows.   module Foo     ITEM_1 = "item1"   end =>   class Foo < ::String     ITEM_1 = Foo.new("item1"   end
Amazon EC sample did not work.  element definition in an Array definition wrongly defined as a type.  fixed.
closes #187, #188.


git-svn-id: http://dev.ctor.org/svn/soap4r/trunk@1679 23c4dcf6-64ea-0310-84d7-d424c7495ec0

+++ test/wsdl/rpc/test_rpc.rb
@@ -49,7 +49,7 @@ class TestRPC < Test::Unit::TestCase
 
   def teardown
     teardown_server
-    File.unlink(pathname('echo.rb'))
+    File.unlink(pathname('echo.rb')) unless $DEBUG
     @client.reset_stream if @client
   end
 
@@ -94,19 +94,19 @@ class TestRPC < Test::Unit::TestCase
     @client.endpoint_url = "http://localhost:#{Port}/"
     @client.wiredump_dev = STDOUT if $DEBUG
 
-    ret = @client.echo(Person.new("normal", "", 12), Person.new("Hi", "Na", 21))
+    ret = @client.echo(Person.new("normal", "", 12, Gender::F), Person.new("Hi", "Na", 21, Gender::M))
     assert_equal(Person, ret.class)
     assert_equal("Hi", ret.family_name)
     assert_equal("Na", ret.given_name)
     assert_equal(21, ret.age)
 
-    ret = @client.echo(Person.new("dummy", "", 12), Person.new("Hi", "Na", 21))
+    ret = @client.echo(Person.new("dummy", "", 12, Gender::F), Person.new("Hi", "Na", 21, Gender::M))
     assert_equal(Person, ret.class)
     assert_equal("family-name", ret.family_name)
     assert_equal("given_name", ret.given_name)
     assert_equal(nil, ret.age)
 
-    ret = @client.echo_err(Person.new("Na", "Hi"), Person.new("Hi", "Na"))
+    ret = @client.echo_err(Person.new("Na", "Hi", nil, Gender::F), Person.new("Hi", "Na", nil, Gender::M))
     assert_equal(Person, ret.class)
     assert_equal("58", ret.given_name)
     assert_equal(nil, ret.family_name)
